Revamped the design in terms of the IPv8 based Peers. The design should now provide a greater degree of flexibility for writing test cases.
