+++ ipv8/test/REST/rtest/peer_communication.py
@@ -1,6 +1,5 @@
 
 from abc import abstractmethod
-from twisted.internet.defer import inlineCallbacks
 
 
 class PostStyleRequests:
@@ -9,7 +8,6 @@ class PostStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_attestation_request(self, param_dict):
         """
         Generate an attestation request
@@ -20,7 +18,6 @@ class PostStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_attest(self, param_dict):
         """
         Generate an attestation
@@ -31,7 +28,6 @@ class PostStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_verify(self, param_dict):
         """
         Generate an attestation verification request
@@ -48,7 +44,6 @@ class GetStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_outstanding(self, param_dict):
         """
         Generate a request asking for the outstanding attestation requests
@@ -59,7 +54,6 @@ class GetStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_verification_output(self, param_dict):
         """
         Generate a request asking for the outputs of the verification processes
@@ -70,7 +64,6 @@ class GetStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_peers(self, param_dict):
         """
         Generate a request asking for the known peers in the network
@@ -81,7 +74,6 @@ class GetStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_attributes(self, param_dict):
         """
         Generate a request asking for the known peers in the network
@@ -92,7 +84,6 @@ class GetStyleRequests:
         pass
 
     @abstractmethod
-    @inlineCallbacks
     def make_drop_identity(self, param_dict):
         """
         Generate a request for dropping an identity
